actually use push target filters when new books are published

+++ api/controllers/BooksController.js
@@ -7,6 +7,7 @@
 
 const HttpError = require('../errors/HttpError')
 const request = require('request')
+const uriRegex = /^(.+:\/\/)?(.+\.)*(.+\.).{1,}(:\d+)?(.+)?/i
 
 module.exports = {
   publish: async function (req, res) {
@@ -23,21 +24,31 @@ module.exports = {
       if (bookExists) {
         throw new HttpError(400, 'Version already exists')
       } else {
-        result = await Book.create(body).fetch()
+        const { title, isbn, author, publisher } = body
+        // require at least 2 fields to be filled out
+        if ([title, isbn, author, publisher].reduce((a, x) => a + (x ? 1 : 0), 0) >= 2) {
+          result = await Book.create(body).fetch()
+        } else {
+          throw new HttpError(400, 'Please fill out at least 2 fields (title, author, publisher, isbn)')
+        }
       }
 
-      req.file('opds').upload(sails.config.skipperConfig, async function (err, uploaded) {
-        if (err) {
-          await Book.destroy({ id: result.id })
-          throw new HttpError(500, err.message)
-        }
-        const fd = (uploaded[0] || {}).fd
-        await Book.update({ id: result.id }, { storage: fd })
-        sendUpdatesAsync(result.id)
-        return res.json({
-          ...result
+      if (req.file('opds')) {
+        req.file('opds').upload(sails.config.skipperConfig, async function (err, uploaded) {
+          if (err) {
+            await Book.destroy({ id: result.id })
+            throw new HttpError(500, err.message)
+          }
+          const fd = (uploaded[0] || {}).fd
+          await Book.update({ id: result.id }, { storage: fd })
+          sendUpdatesAsync(result.id)
+          return res.json({
+            ...result
+          })
         })
-      })
+      } else {
+        throw new HttpError(400, 'Missing OPDS file upload')
+      }
     } catch (e) {
       if (e instanceof HttpError) return e.send(res)
       return res.status(500).json({
@@ -73,16 +84,27 @@ module.exports = {
 async function sendUpdatesAsync (id) {
   const book = await Book.findOne({ id })
   const targets = await TargetUrl.find()
+  if (!book) return
   for (const i in targets) {
-    sails.log('sending ' + book.id + ' info to ' + targets[i].url)
-    request.post({
-      url: targets[i].url,
-      headers: { 'User-Agent': 'RoE-aggregator' },
-      form: book
-    }, function (err, httpResp, body) {
-      if (err) {
-        sails.log(`error: failed to send book ${id} to ${targets[i].url}`)
-      }
-    })
+    const item = targets[i]
+    const { author: fAuthor, publisher: fPublisher, title: fTitle, isbn: fIsbn, url } = item
+    const { author: bAuthor, publisher: bPublisher, title: bTitle, isbn: bIsbn } = book
+    sails.log('sending ' + book.id + ' info to ' + url)
+
+    if (uriRegex.test(url)) {
+      if (fAuthor && !((bAuthor || '').includes(fAuthor))) continue
+      if (fPublisher && !((bPublisher || '').includes(fPublisher))) continue
+      if (fTitle && !((bTitle || '').includes(fTitle))) continue
+      if (fIsbn && !((bIsbn || '').includes(fIsbn))) continue
+      request.post({
+        url: url,
+        headers: { 'User-Agent': 'RoE-aggregator' },
+        form: book
+      }, function (err, httpResp, body) {
+        if (err) {
+          sails.log(`error: failed to send book ${id} to ${url}`)
+        }
+      })
+    }
   }
 }
